+++ counter/client/lib/main.dart
@@ -10,7 +10,7 @@ void main() {
 
   final projectId = 'd2sqf8kgc98s73838big';
   final apiKey =
-      'sk-client-43b6916d2e6f733bHSaXZP17EP9Jyf5WxrtUPRzRxyx_-FQXmbp8hn7548Oouj3QC8Em65hYY1c8MJTtl2JuZfQBLwJIUZI_S-vg7w==';
+      'sk-client-UCcOxMn4MIKXTY_BBCvronXvGrJXEvbeedH6wR30EHlZ0Uh8Q_LUaOlM3vReWPBTZd5wm8NXQdmVlGQCVBTIWA';
 
   final url = 'wss://api.horda.dev/$projectId/client';
   // final url = 'ws://localhost:8080/client'; // For local development