+++ twitter/client/lib/config.dart
@@ -1,6 +1,6 @@
 const kProjectId = 'd368c1sgc98s738ue7cg';
 const kApiKey =
-    'sk-client-a7e1df6be79bb12dn8gWLaJt0T9HhkLrCf48tpoD65atXIJyYDegp4DxJcikSbzXi5GaztaI7RoN2eLvxzJ6gRfLFuwkqkJtV_0D3Q==';
+    'sk-client-1Q02wJvUuZkkDMCwlnlGdmdKbJYP8fdQ9p0P5Wq8PLv4XKLlPN-UGeHPU0nsQZqYuDooX6XX9sMyjy1UoSuGYg';
 
 final kUrl = 'wss://api.horda.dev/$kProjectId/client';
 // final kUrl = 'ws://localhost:8080/client'; // For local development